refactor(p2p): isolate protocol admission adapter
